# backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Evaluation API
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local backend server

paths:
  /vehicles:
    get:
      summary: Get all vehicles
      operationId: getVehicles
      tags:
        - Vehicles
      responses:
        "200":
          description: Successful response with HAL-style _embedded
          content:
            application/json:
              schema:
                type: object
                properties:
                  _embedded:
                    type: object
                    properties:
                      vehicles:
                        type: array
                        items:
                          $ref: "#/components/schemas/VehicleDto"
                  _links:
                    $ref: "#/components/schemas/Links"
                required:
                  - _embedded
                  - _links

  /vehicles/{id}:
    get:
      summary: Get vehicle by ID
      operationId: getVehicleById
      tags:
        - Vehicles
      parameters:
        - name: id
          in: path
          description: ID of the vehicle to retrieve
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Vehicle found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleDto"
        "404":
          description: Vehicle not found

    delete:
      summary: Delete a vehicle by ID
      operationId: deleteVehicle
      tags:
        - Vehicles
      parameters:
        - name: id
          in: path
          description: ID of the vehicle to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Vehicle deleted successfully (No Content)
        "404":
          description: Vehicle not found

components:
  schemas:
    Link:
      type: object
      properties:
        href:
          type: string
      required:
        - href

    Links:
      type: object
      properties:
        self:
          $ref: "#/components/schemas/Link"
        allVehicles:
          $ref: "#/components/schemas/Link"
        getVehicleById:
          $ref: "#/components/schemas/Link"
        delete:
          $ref: "#/components/schemas/Link"

    VehicleDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        vehicleType:
          type: string
        created:
          type: string
        mileage:
          type: string
        gearbox:
          type: string
        owner:
          type: integer
          format: int64
        kwAndPs:
          type: string
        hek:
          type: number
          format: double
        taxation:
          type: string
        _links:
          $ref: "#/components/schemas/Links"
      required:
        - id
        - vehicleType
        - created
        - mileage
        - gearbox
        - owner
        - kwAndPs
        - hek
        - taxation